Allow the S-parameter analysis to handle DC and negative frequencies
The S-parameter analysis generates frequency domain small signal
parameters which can then be used in the S-parameter model. The S-parameter
model can be used in frequency domain simulation, such as HB. It will
be helpful to allow SP analysis to handle DC and negative frequencies so
the data can be directly used in the HB simulation. This commit make it
possible to handle DC and negative frequencies in the linear type of frequency
sweep. This is also compatible with other simulators, such as Hspice. The
results match Hspice result