Refine Trlinos probing messages and add CURL probe
This doesn't significantly change the way Trilinos is probed, but cleans
up the messages. It also adds some informational probes for the user
(that apply to optional linear solver features of Xyce).

I also added an inital probe for CURL, using the default CMake probe. It
may need to be refined over time.