Make allocateVoltageLimitingSolver return a value
This function is declared as returning a bool but returns nothing.  I have
found *EXACTLY* one system (my home BSD system, but not my almost identical
office BSD system) where this leads to segfaults.  It impacts NO supported
builds.

I am making it return true just so that when we do a push to github on Monday,
it'll be correct everywhere, not just on supported builds.

There were other instances of the same problem that I found last week, but
Eric already fixed those.