Make allocateVoltageLimitingSolver a void
In my last commit, I made allocateVoltageLimitingSolver return true,
because it was declared as a bool and had previously returned nothing.
But when it's called, it doesn't actually have its return value used
at all, so the right fix would be to just make it a void function,
like Eric had done for the other instances of this problem.