Change Default for .OPTIONS OUTPUT PHASE_OUTPUT_RADIANS
This commit changes .OPTIONS OUTPUT PHASE_OUTPUT_RADIANS=<boolean>
so that its default value is now false.  This means that the default
output for phase output, for the VP(), IP(), SP(), XP() and ZP()
operators and for .PRINT SENS for a .AC analysis, is now in degrees.
The phase angles in Touchstone files are still only output in degrees,
per that standard.

SON Bug 1208
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1208