Clean up cmake-refactor to match master branch
A few things got missed in the updating of this branch. Some missing
files in the various CMakeLists.txt, a missed file deletion, and some
small whitespace changes. Some of the source files also had some small
changes (which I think are remnants of the original refactor effort).
This commit should make everything synchronized, and up-to-date.