Cleanup and hack in Intel MKL capability
The Find MKL script does not work for our default system. The lore is that it
is notoriously difficult to have CMake properly handle the Intel MKL. Just to
get Xyce to compile, I hacked in the Intel 18 compile and link lines directly,
and put in a (commented-out-by-default) target_link_libraries line with the
proper information. Note that in the current system, all this is needed only
for the FFT library. (The MKL also also includes access to the Pardiso solver,
but that would be via Trilinos. It would require a different
check/configuration.)

In addition, I made the required Trilinos version 12.12, because the reporting
version on CEE is 12.12.0, for some strange reason. (Official releases always
start with a X.X.1, not a X.X.0.)

Finally, I fixed a lot of messages that should be labeled "STATUS".