Remove false promises from release notes
There are several features that we are saying are deprecated (i.e.,
"don't use these anymore, even though they might still work") with
bold claims that they will be removed at a specific time.  Those
specific times have come and gone without the removal of those
features.

Instead of saying when they will be removed, we now say that they are
deprecated, what to use instead, and that the deprecated feature may
be removed without notice at some time in the future.  That should
cover the bases.