Change warning on automatic double quote to FILE conversion
The use of double quoted filenames in model parameters are
depreciated and the warning said that this would be removed
in Xyce 6.11.  That didn't happen but it will be taken out
soon. So the warning is now that this is deprecated and
will be removed in a future version of Xyce