Fix up CMake FFT search, and some cleanup
The FFT search now follows the proper logic of using the Intel MKL, if it's
available, or, if not, searching for FFTW. It also sets the right flags for
Xyce. If using the Intel MKL, this setup assumes that the user has properly set
all the flags in the Trilinos build, such that linking works. This is not
trivial, and relies on the user to know what they are doing. I have relied on
the Intel MKL Link Line Advisor:
<https://software.intel.com/en-us/articles/intel-mkl-link-line-advisor>.

In this commit, I also removed the setting of CXX and C flags based on the
Trilinos flags.  Originally, I thought this would be the mechanism for handling
the MKL library flags, but I was incorrect.  Now it makes sense to use
CMAKE_BUILD_TYPE (or some equivalent method).

Finally I updated some status messages (making them clearer and/or more
CMake-like).