Add most missing features to CMake
This commit adds a lot of the optional flags that are needed for Xyce to
operate to its full potential. That includes big modifications to the .h
file template. I also removed the "defaultConfig.cmake" file, and moved
the relevant features to "simple_features.cmake". This commit also
cleans up some logic and modifies some comments.