Fixes supernoding algorithm to perform correctly in parallel
Increasing the number of processors used to simulate the supernoding tests
illustrates that there are issues of incorrectness and code crashing.
Among the issues with the supernoding code that has been in Xyce for some time is:

1) Lack of checking for whether the local node being removed or the local node
being replaced is actually on this processor.  The global supernode list might
cause a processor to remove a local node for a node that is not currently on
the processor.  Furthermore, the global supernode list might cause the processor
to look for a local node that is being removed that may not be on that processor.
Either way, the code did not gracefully handle these conditions, which resulted
in illogical behavior or seg faults.

2) Checking for nodes being removed on other processors that are not in the
local supernode list.  The distribution of the devices might be such that a
processor does not have a supernode list, but there are nodes being removed
on other processors that are also local to this processor.  This case was
not being addressed in the current implementation.

3) Harden node removal code in templated graph class to inquiries related
to (1).  Some basic graph inquiries necessary in the implementation of (1)
were not hardened to the situation where the node in the graph did not exist.

This addresses the SON bug 1246.