Adds note for direct solver performance improvements / memory reductions
The recent change to remove the ground node from the graph and matrix has
reduced the number of graphs necessary for Xyce to execute in serial and
parallel.  In serial, the ground node removal has eliminated the need for
the additional matrix copy for the direct solvers, thus improving performance.