Moves extractParamData and extractGlobalParamData to parsing helpers
The code that extracts data from .PARAM and .GLOBAL_PARAM statements
is being moved back into the IO package from the Device package.  This
makes the processing of '.' lines a little more transparent and reduces
the pollution of one package to another.

Now handleLinePass1 traps for '.PARAM' and '.GLOBAL_PARAM' and calls
the extractParamData and extractGlobalParamData, respectively, to process
those lines.  The way this was done previously was through the extractData
call that uses the options package to process the '.' lines.  So, the
net effect of this change is relatively nil.  This is because the options
package would use the extractData and extractGlobalParamData in the device
package to just add the parameter or global parameter to the CircuitBlock
object it was passed.