Removes copy constructor from circuit context object
This object does not need a copy constructor.  In parallel the unpacking
of the object can be done directly in the object.  It is not necessary to
use an intermediary object and then copy it back into the circuit context.

This clarifies where the circuit context is resolved in parallel.