Adds two tests to work better with Windows issues.
The first issue is that unistd.h is not normally present on a windows
system. Although Xyce requires cygwin for its tests which does include
it, I went through the process of getting xyce's code to compile without
it.
The second issue is that Trilinos+Xyce cause a bazillion sprintf
security warnings when compiling Xyce. This is annoying because the
security warnings boil down to privilege escalation through a potential
buffer exploit. But xyce isn't a privileged program or one with any
networking, so, long story short the user who runs Xyce already has all
the privileges they could get out of hijacking xyce's run.