start to add content to the Ylin device
The Ylin device is going to be a general frequency domain device
that can use the frequency domain small signal parameters, such as
Y, Z and S parameters. This commit adds the Jacobian stamp using real
information from device instance, corrects the device traits, removes
unnecessary functions. The Ylin device is not a linear device in
time domain for the device evaulation purpose. But it is a linear device
in frequency domain. It can be used in both time and frequency domain