Change/add CMake probes for Trilinos features
First, I changed the way CMake looked for required Trilinos features.
The CMake will now look for all the features, and then thow a fatal
error.  This will allow a user to see all the missing flags that have to
be added to the Trilinos CMake invocation to make it work for Xyce.

I also added probes for Amesos2, Stokhos, and ROL. As for ShyLU and
Basker, I haven't yet created a Trilinos build with those enabled, so
I'm not sure if I'm probing for the right thing.  This is particularly
true for Basker, as we aren't even sure which Basker to enable.