Remove comments about "EXPERIMENTAL" Xyce/ADMS
Back when I was still working on bug 1168 I had all of my changes on a
private branch, but the modified Xyce/ADMS back-end was in a state
where *SOME* models could be processed with it.  During that time I
had the "experimental" back-end emit comments into the code it output
warning that the code should never be regenerated with the officially
released version, lest the optimizations be lost.

When I actually got bug 1168 to the point where I was comfortable
making it the officially released version and the old one the "legacy"
version, I forgot to remove these warning comments.  I noticed today.

I would have liked to make the 6.12 release not have these comments,
but it is really unlikely that anyone outside the labs will mess with
the code ADMS produces, and so it is unlikely anyone will notice the
comments.  In the meantime, I'm getting rid of them on the development
branch.