Undo fix of a "typo" that was not a typo
In the release notes, I had deliberately refered to a long-standing
bug in ADMS's implicit rules using the present tense, because the bug
is in fact still there.  The fix of bug 1223 merely works around that
issue.  It does this by telling ADMS not to use its implicit rules,
and to use my modified stuff explicitly instead.

I did feed the fixes to the implicit rules back to ADMS developers as
a github pull request, but that pull request has not been accepted
yet.

My text got rewritten into the past tense as a cleanup of "typos", and
in fact changed the meaning, implying that ADMS has been fixed.  ADMS
still has this bug, meaning it still incorrectly propagates
dependencies and leads to incorrect code generation if those implicit
rulese are used.  I did not fix ADMS, I only worked around it.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1223