Support I(*) for BJTs and FETs
This commit adds support for ID(*), IG(*) and IS(*) for JFETs,
MESFETs and MOSFETs.  It also adds support for IB(*), IC(*) and
IE(*) for BJTs.  More complex wildcards, like I(X1*), are still
only supported for two-terminal devices.

As a final note, I(*) will print none of the lead currents for the
J, M, Q and Z devices.  This behavior differs from HSPICE.

SON Bug 983
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=983