Minimal CPack and .so builds work
the issue preventing building a dynamically shared library was that
ksparse also defines a main entry point and the linker didn't like that.
ksparse/driver.c is no longer in the sources list.
For CPack, I've enabled a very minimal but functional configuration.
I tried to use the more complex Xyce version number but RPM refuses colons in
file names, so that is a bit problematic.