Add frequency domain model for S-parameter devices
The frequency domain S-parameter model is implemented and it can
be used in frequency domain simulations, such as HB. The frequency
domain vector and matrix loading functions are added.

The frequency domain S-parameter model is tested by transmission
line examples. The S-parameter analysis is used to generate small
signal parameters for these transmission lines. These parameters
are then read in as S-parameter models and the HB simulations are
performed. The results from original transmission lines (with thousands
of components) are compared to the results from S-parameter models
(with only a few components). They match well for HB