Change How V(), I() and P() Handle * in Node/Device Names
This commit changes how the V(), I(), P() and W() operators handle
the use of the star (*) character in node and device names on .PRINT
lines.  Previously, something like V(1*) would yield an explicit match
for node 1* if that node existed in the netlist.  Now, for example,
V(1*) is solely a wildcard request for all node names that start
with '1'.  This matches the HSPICE behavior.

The changes in printLineDiagnostics() make the searches for node or
device names slightly more efficient.  Most of the changes to
removeStarVariables() are just indentation changes.  The substantive
change was removing the check for a node/device name that explicitly
matches the wildcard request (e.g., for node 1* for the wildcard
request V(1*)).

As a final note, .PRINT {V(1*)} {I(R1*)} will still look for explicit
matches for node 1* and device R1*.  Wildcard expansion does not
apply within expression delimiters.

SON Bug 1211
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1211