Fix Ylin device when S and Z parameters are used
The Ylin device uses Y parameters internally. If S or Z parameters
are used in the touchstone file, they are converted to Y parameters.
This is done during the touchstone file parsing. However, the Y
parameters are not shaped to the correct size before the conversion.
This leads to seg fault when it is used during the device loading.
It cannot be tested before the implementation of the Ylin device.
This commit fixes the problem