Support IB(*) for MOSFETs
This commit adds support for IB(*) for MOSFETs, which is the bulk
lead current.  Previously, IB(*) was only supported for the base
lead current for BJTs.

This commit's approach is a bit inefficient, since it stores the
MOSFET device names in one set, that is used for IB(*).  It then
stores all of the FET names (for JFET, MESFET and MOSFET) in a
second set that is used for IS(*), IG(*) and ID(*).  This might be
refactored at some point, when the issue of the ordering of the
wildcard requests in the output file is addressed.

This commit also re-named most of the variables in the "marshall'ing
loop" in removeStarVariables(), in addition to adding handling of the
separate lists for IB(*) for BJTs and MOSFETs to that loop.  That
renaming makes it easier to re-name/add variable lists in that loop.

SON Bug 983
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=983