Add SON Bug 1208 to Xyce 7.0 Documentation
This commit adds information on SON Bug 1208 to the Xyce 7.0 Users'
Guide, Reference Guide, and Release Notes.  That bug was about the
change to degrees (from radians) as the default units for the Xyce
phase operators (e.g., VP()) and AC sensitivity output.  A note about
the new .OPTIONS OUTPUT PHASE OUTPUT RADIANS=<boolean> was added to
the .OPTIONS section of the Reference Guide.

SON Bug 1208
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1208