Add SON Bug 1252 to Xyce 7.0 Release Notes
This commit adds SON Bug 1252 to the Fixed Defects Table in the
Xyce 7.0 Release Notes.  That bug was about making .OPTIONS OUTPUT
PRINTHEADER=FALSE work with the outputters for .PRINT SENS for an
AC analysis, .PRINT ES, and .PRINT PCE.  The release note entry
does not mention .PRINT PCE, since PCE is not documented in the
Reference or User's Guide yet.

SON Bug 1252
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1252