Improve the frequency domain S-parameter models
The initial frequency domain S-parameter device uses the data at
exact frequencies requested by HB simulation. This is not very
practical. This commit adds linear interpolation/extrapolation to
handle the practical data from touchstone file in HB simulation.
The initial results show good agreement between the original
circuit and the circuit with S-paramter devices for transmission
line applications.

The S-parameter is a linear device only for frequency domain device
evaluation. However, it is a nonlinear device for time domain
evaluation. This is different to many other "linear" devices in Xyce.
For HB simulation, the S-parameter was loaded in both frequency and
time domain, which is wrong. The time domain evaulation has not been
implemented. It behaves like a resistor. But this will change the
circuit that uses S-parameter device for HB simulation. The previous
examples work due to the specific way the S-parameter device is connected.
It did not change the voltage at the node, but it changed the current.
This commit fixes the linear/nonlinear and time domain/frequency domain
device evaluation for S-parameter model. The device is evaluated either
in frequency domain or as nonlinear device in time domain. The currents
now match between the original and S-parameter HB runs