Document -quiet Command Line Option
This commits adds the -quiet command line option to the Command
Line Arguments tables in the Reference Guide and Users' Guide.
It also makes other formatting/wording changes to make those two
tables more similar in content.

This commit also removes -info from the Command Line Arguments
table in the Users Guide.  That argument was removed in Xyce 6.7,
as part of SON Bug 838.  We missed updating the Users' Guide table,
at that point.

SON Bug 891 and 838
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=891
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=838