Adds .OPTIONS LINSOL-AC to user and reference guide
This commit includes documentation of the .OPTIONS LINSOL-AC option
that has been in the code.  It allows AC analysis to use any of the
transient linear solver methods, since the linear system is real-valued
due to the use of real equivalent forms.  Previously, the linear solver
was hard coded to use KLU.  This is not always desirable if there are
better options for direct or iterative methods.