Removes static integer first_time from Ksparse solver
This static integer is set after the first symbolic analysis and
will keep any future symbolic analysis from being performed, regardless
of whether the solver object is destroyed and recreated.  This is
the case when Xyce uses a sequence of analysis types, like DC and AC
or DC and HB.  It is particularly egregious because the two sequential
analyses, like the examples stated, generate different sizes of linear
systems.  Therefore, the symbolic analysis for the DC phase is irrelevant
for the following analyses.  The confusion affects the solver
in various ways, like tolerances not being set, and results in different
behavior from when it is used only in one analysis.

The same integer, indicating that this is the first time the solve is
being performed, is maintained in the interface between KSparse and
Xyce.  This integer should be used instead because it is correct.

This commit addresses SON bug # 1255