Remove outdated, erroneous comments
There were some comments in the xyceAnalogFunction_nosac.xml file that
suggested it wasn't finished and that analog function calls that had
output arguments would not be handled correctly.  Those comments were
true when they were written, but were there to remind me to write the
code needed for that case.  I wrote that code long before committing
the files to the repo, but never removed the comments.