Fix mishandling of analog function
An open source user pointed out that Xyce/ADMS threw an error message
when he tried to call an analog function that took both input and
output arguments, but only if he used "$vt" as  one of the input
arguments.

The reason for this was a careless clearing of a variable outside any
conditional when that variable had only ever been SET inside a
conditional.

Since the only use of the variable outside the conditional was to
clear it, there is no harm SETTING it outside the conditional, too.

This fixes the user's use case and has no impact whatsoever on any of
the models already in Xyce.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1256