Document bug 1256 fix
I fixed a Xyce/ADMS bug and am adding it to the fixed defect table.

While I was at it, I took the opportunity to remove some "this is
fixed now" text from another entry.  We have been in the habit of
adding this text to every entry for years (to the point that sometimes
people have even considered its omission a mistake), but it is really
unnecessary.  This is the *table*of*fixed*defects*, which has "Fixed
Defects" as its caption and its title.  Simply saying that a defect
enumerated here is fixed is redundant.  It's high time we get out of
this habit (which is as hard to kick for me as it is for anyone).

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1256