Fix error in assert text
An assert had the special string "$(.)" in it, when it should have had
"%(.)".  This meant that when the assert failed you got an error about
the assert text instead of the assert's text.

This assert is tripped when a model tries to do more than one noise
function in a single contribution, as EKV 2.6 does (there's nothing wrong
with doing that, it's just that at the moment Xyce/ADMS doesn't know
how to handle it, so the contribution needs to be split into pieces).
It is good for the assert to emit its error, and bad for ADMS to emit
a different error instead.