Fix SON Bug 928 (HB_IC Output with .STEP)
This commit provides a fix for SON Bug 928, which was about HB_IC
output for the .STEP case.  The approach is, for the FORMAT=STD case,
to use a <netlistName>.hb_ic.prn.tmp file to capture each transient
IC simulation. The contents of <netlistName>.hb_ic.prn.tmp is then
copied to the end of <netlistName>.hb_ic.prn output file, when
HB::runTol() accepts an IC run for a given STEP.  The temporary
file <netlistName>.hb_ic.prn.tmp is then deleted.  This process
is repeated for each STEP iteration, so that only the results from
the accepted IC runs appear in <netlistName>.hb_ic.prn.

In order to avoid muddying up the existing TRAN outputters, new
outputter classes were made for the HB_IC outputters.

SON Bug 928
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=928