Add standard includes to new files
Three new outputters make use of ifstreams, but did not explicitly
include the headers needed to access the full ifstream objects (they
did, however, at least get forward declares from somewhere).  This
broke Charon's build, but not our own.

My speculation is that we were somehow picking up an indirect include
of these standard headers through Trilinos.  We use an old version of
Trilinos in nightly builds, but Charon uses trilinos off of the
development branch.  My guess is that Trilinos cleaned out some
unneeded header includes.

It is best practice to include standard headers you need explicitly in
code, and not rely on them being brought in indirectly.  So even if
this doesn't fix Charon's build, it is the right thing to do.