Begin to refactor analog function derivatives
Per bug 1253 on joseki, our code for handling analog function
derivatives without Sacado is extraordinarily inefficient, as it
(effectively) causes the function to be called repeatedly every
time its derivatives with respect to probes are needed, despite all
the internal calculations of the function not actually depending
directly on any probes (indirectly, yes).

This commit is the beginning of an attempt to rectify that, by making
analog function calls be precomputed for assignments just as builtin
functions are.  To do this requires a little finesse.

Instead of simply emitting calls to the templated analog function to
get the value, and repeatedly calling d_functionName for every probe
the assignment depends on, we'll create a functionNameEvaluator object
whose constructor will actually evaluate the function and its
derivatives with respect to its arguments.  Then, when we actually
need the function value we'll emit getValues and getDerivs calls on
that Evaluator object to retrieve precomputed values in the former
case, and to apply the chain rule to precomputed derivativesin the
latter.

This should dramatically improve performance of codes that offload a
lot of their computation to analog functions.

This commit merely creates the declarations of the classes and
implementations of some of their methods.  It does not emit the actual
evaluator function yet, but that is a tiny delta off of of what's
done in the current d_functionName functions, and I expect it to be
done in the next commit.

Nothing in this commit actually *USES* the classes and methods we're
producing yet.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1253