Remove Bug SON 652 (HB Output Bugs) from 7.0 Rel. Notes
This commit removes SON Bug 652 from the Known Defects table in the
Xyce 7.0 Release Notes.  That bug was the umbrella bug for HB outputter
bugs.  I decided to not include that bug in the Fixed Defects table
since it was an umbrella bug.  That decision can be changed during
bug QA.

SON Bug 652
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=652