Add "evaluator_" methods to analog function classes
This commit adds the guts of the evaluator_ method to the classes
created by the previous commit on this branch.  The constructor for
the class takes all the same arguments as the analog function itself,
and then calls the evaluator_.  The return values and their
derivatives with respect to input arguments are computed and archived
as member variables of the class objects.  The getValues and getDerivs
methods simply make use of these precomputed values.

This commit produces code that compiles, but none of the new code
actually gets used yet.  That's the next step.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1253