Make analog functions be precomputed
Xyce 6.12, like ngspice and qucs, and any other codes that use ADMS
templates that have their roots in code Laurent Lemaitre wrote,
precomputes builtin functions and their derivatives when these
functions appear on the RHS of an assignment.  For example, the
verilog-a line
    variable = max(exp1,exp2);

gets expanded into something like:

  {
     value_max_0 = max(exp1,exp2);
     deriv_max_0_d0 = (exp1>=exp2)?1.0:0.0;
     deriv_max_0_d1 = (exp1>=exp2)?0.0:1.0;

     d_variable_dfoo = deriv_max_0_d0*dexp1_dfoo + deriv_max_0_d1*dexp2_dfoo;
     d_variable_dbar = deriv_max_0_d0*dexp1_dbar + deriv_max_0_d1*dexp2_dbar;
     variable = value_max_0;
   }

That is, the function and its derivatives with respect to its
arguments are precomputed, and the precomputed values used wherever
the function or its derivatives with respect to a desired independent
variable are needed.  ADMS keeps a little datastructure of what
functions appear on the RHS for precisely this reason.

Many other codes share a common ancestor of ADMS back end, and all of
them do this sort of precomputation.  But the precomputation is only
done on builtin functions.  Analog functions are handled very
differently, and these are NOT included in the ADMS data structure
that lists the functions called in an expression.  When I refactored
Xyce/ADMS to rely less on Sacado, I copied this approach.

Turns out that the approach is extraordinarily inefficient for analog
functions, where the back-end would generate a function func and a
function d_func for each analog function. func(args) would return the
regular analog function value, and d_func(args,dargs) would return the
total derivative of func with respect to a desired independent
variable.  d_func is very much like func, but it also computes all the
derivatives of all its internal variables with respect to all of its
arguments, then applies the chain rule using "dargs"(the derivatives
of the arguments with respect to the desired independent variable).

This approach ends up calling a function even more complex than the
original function several times every time the function appears in an
expression.  For models that make heavy use of analog functions this
can end up dominating the residual load.

This commit changes adms.implicit.xml to add analog function calls to
the "function" list that used only to contain builtins, and then use
this information to precompute the the function and its derivatives
with respect to its arguments using the "Evaluator" classes that were
added in the previous two commits on this branch.

This change squeezes even more performance out of BSIM-CMG on top of the
improvements that were obtained from pitching Sacado, and should make
models that rely heavily on analog functions even more efficient.

Note that this commit does not yet complete work on bug 1253, as
analog functions themselves do not do function precomputation
internally, even on builtins.  So a further improvement is necessary
to extend the precomputation approach to analog function bodies.  This
commit does NOT do that part.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1253