Assure we are using consistent header XML
Until the analog function refactor for bug 1253, the header templates
for sacadoized and non-sacadoized versions weren't very different, and
if a code wasn't using analog functions just so, it didn't matter
which you used.  But it's now wrong to use the one without "_nosac" in
its name with the non-sacado back end.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1253