Hack "hypot" support into the legacy ADMS back-end
When I added hypot support to Xyce/ADMS, it was after I had declared
the Sacado version "legacy" --- and therefore I had only done it in
the new back-end.

I am now currently tangled in figuring out why a user is claiming that
the old back-end generated faster code for his model than the new.
His model uses hypot, so in order to do comparisons I need this in the
legacy back-end, too.