Removes unused information and changes const members to non-const
The augmentStart() and augmentLoc_ method and pointer, respectively, is not
properly set by any of the constructors in the BlockVector class and is not
used.  So, it is being removed before is confuses anyone.

Some of the private integers in the BlockVector were constant integers, which
doesn't make sense.  In the case of an augmented vector, these integers may
need to be recomputed, so they are not constant.

One of the commonly used constructors for augmented BlockVectors in HB has
been corrected to recompute several local integers based on the number of
augmented rows.  This was not the issue in bug 1263 (SON), but it still is
the right modification to make.

Bug#: 1263 SON