Changes storage of augmented LIDs for overlap block map
The overwritting of the augmentedLIDs_ by the overlap block map
constructor resulted in the wrong LIDs being used by the oscillator
HB algorithm.  The changed the performance of the algorithm and
also caused a seg fault.  The overlap augmented LIDs are now stored
in a second vector for use later, if necessary.

Bug#: 1263 SON