Add Mixed Signal Methods to Get Device Param Values
This commit adds the checkDeviceParamName() and getDeviceParamVal()
methods to the Python interface to the Mixed Signal interface.
The underlying XyceCInterface methods are xyce_checkDeviceParamName()
and xyce_getDeviceParamVal().  These methods allow the user to query
for device parameter values using the fully-qualified name for the
parameter (e.g., X1:R1:R for device R1 in subcircuit X1).