Import original BSIMCMG 108 files
As a first step in implementing BSIM-CMG 108 in Xyce, I'm importing
the original files from the BSIM group.  These are not ready to use
and need to be "Xycified", but I want the changes to be clearly
recorded in git history as well as the README_Xyce I'll write
afterwards.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1258