Get BSIM-CMG 108 "Xycified" enough to compile
This commit makes changes necessary to get BSIM-CMG 108 compilable as
a plugin for Xyce.  Changes that had to be made to the source are
explained in README_Xyce.

This is not yet wired in to Xyce proper, nor is it tested beyond
compilability.

BSIM-CMG 108 does not have multiplicity kludged in as BSIM-CMG 110
does, and expects that the simulator takes care of that implicitly.
Since we are going to need multiplicity to work for the PDK we are
trying to support, I will have to hack that in myself.  That is NOT
done by this commit.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1258