Fix serious convergence problems
BSIM108 introduced a new correlated thermal noise model that uses an
internal node N.  If the thermal noise parameter TNOIMOD is not set
to 2 this model isn't used and the N node floats.  The result is
guaranteed convergence failure.

BSIM-CMG 110 fixed that bug by making sure to collapse N to ground
when it isn't needed.  It took me a while to figure this out, because
so many other changes were made between 108 and 110 that it was hard
to spot this minor difference.

Once I back-ported the correct node collapse, the convergence problems
went away.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1258