Add Mixed Signal Methods to get GIDs of Adj. Nodes
This commit adds the getNumAdjNodesForDevice() and getAdjGIDsForDevice()
methods to the Python interface to the Mixed Signal Interface.  The
user can now request the GIDs (of voltage nodes and the ground node)
that are adjacent to a specified device.  This may be helpful when
using the Xyce parser to extract the graph topology of a netlist, for
subsequent use in an external graph-analysis package.

SRN Bug 1052
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1052