Change placeholder version to 8.0
For years now our development branch has had "7.0" coded in its
configuration as a placeholder for the real version number (this
placeholder is not ever printed, as development versions get the date
and git SHA-1 instead of the version number).

Since we are on the brink of a real 7.0 release, it's time to have a
different placeholder.