Fix some errors in the BSIM4
Ting had a question about the BSIM4 today, and that got me looking
into bug 807 again.  I am no farther along on understanding what is
wrong with that bug, but in the process of looking over the code I
found some unrelated mistakes that only impact RGATEMOD=3 (which is
not the one with which we're currently having problems).

I've fixed those errors, and also removed some dead relics of early
BSIM4 porting that just confuse matters (comments and unused variables).

Note that while bug 807 is what got me looking at this, this change
has no impact whatsoever on that bug.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=807