Remove small but odd "xyceisms" in bsim4
Going over bug 807, I've had to revisit the implementation of BSIM4 in
Xyce, which was never really completed and has a number of odd relics
of its "old DAE" implementation.  I've gone back to the original SPICE
implementation and am trying to make them match as much as possible,
given the fundamental differences in how Xyce and SPICE models
actually load things under the hood.

In doing so, I'm finding tiny nits to pick, but no real answers to bug
807.  In this commit, I'm removing one such nit, a very strange use of
the store vector to store "von" for no apparent reason.

In the original SPICE implementation, von is both a local variable in
b4ld and a member ("BSIM4von") of the BSIM4 instance structure (the
reason for this is presumably to simplify typing, so that one didn't
have to type "here->BSIM4von" every where it was used --- they copy it
into a local var just called "von".

In Xyce, we've got von as an instance variable in the Instance class,
and a local variable "von_local".  For some bizarre reason we were NOT
using von_local exactly the way that the original implementation
did (copying it out of the struct every time), and were strangely using
the store vector to keep von values from iteration to iteration.

This commit gets rid of that store variable, and makes our use of
von_local be exactly like how they do it in the SPICE implementation
--- as a local copy of the stored von.

von_local is used for voltage limiting purposes, and our strange
misuse of it could have had an effect on our correctness (but didn't
--- if you followed the logic it would have been correct, but it was
really obscured).

Unfortunately this utterly trivial change does absolutely nothing for
the convergence issues of bug 807.  Not much to show for a day of
trying to unwind BSIM4 confusion, but there you have it.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=807