Expose two BSIM-CMG parameters on the instance line
BSIM-CMG has some very clear documentation on which parameters should
be model parameters only (.model card) and which should be instance
parameter only (device instance line), and which should be both (with
the instance specification overriding the model if it is given, and
the model specification being the default otherwise).

A simulator which shall remain nameless violates this routinely,
allowing ALL model parameters to be overridden on the instance line.
And fabs are in the habit of using that simulator to generate and test
their PDKs.

u0mult and delvtrand are two model parameters clearly specified as
model-only in the Berkeley documentation, but which are getting used
in a fab-provided PDK of significance to the team.

Xyce exits with a fatal error if it encounters an instance that uses
undefined parameters.  And this PDK is defining instances that use
non-default values of those two parameters on the instance line, so
Xyce produces incompatible answers if they are just removed.

In order to get past this problem, we're marking these two parameters
as both instance and model.  This is a violation of CMC guidance that
simulators should never change the standard models in non-standard
ways, but it is required to move past this PDK issue.

Under no circumstances should users deliberately use this feature.  It
is provided solely to get PDKs over which we have no control to work
in Xyce.  For this reason, I am NOT documenting this as a feature or
in the reference guide parameter tables just yet.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1269