Add Support for AVG Measure for .AC and .DC Analyses
This commit adds support for the AVG measure for .AC and .DC
analyses.  The FROM-TO qualifiers are supported, but setting
FROM=TO results in a failed measure, as in HSPICE, for the AVG
measure.  Note that Xyce's .MEASURE results differ from HSPICE
for the case where multiple (e.g., N) variables are used on the
.DC line. In that case, Xyce treats the entire .DC analysis as
"one step", while HSPICE treats it as N steps.

SON Bug 1267
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1267